refactor: 진형 관련 패키지를 formation으로 분리하고 입력 매핑 책임을 controller로 이동
- formation 관련 클래스들을 domain.board → domain.board.formation으로 이동
- cho/han 배치 전략을 formation 하위 패키지로 재구성
- PlacementInputMapper를 domain에서 controller로 이동하여 입력 책임 분리
