+++ src/main/java/domain/board/formation/FormationType.java
@@ -1,4 +1,4 @@
-package domain.board;
+package domain.board.formation;
 
 public enum FormationType {
     DEFAULT,
