+++ src/main/java/domain/board/formation/PlacementOption.java
@@ -1,13 +1,13 @@
-package domain.board;
+package domain.board.formation;
 
-import domain.board.cho.DefaultChoPlacement;
-import domain.board.cho.InnerHorseChoPlacement;
-import domain.board.cho.LeftInnerHorseChoPlacement;
-import domain.board.cho.RightInnerHorseChoPlacement;
-import domain.board.han.DefaultHanPlacement;
-import domain.board.han.InnerHorseHanPlacement;
-import domain.board.han.LeftInnerHorseHanPlacement;
-import domain.board.han.RightInnerHorseHanPlacement;
+import domain.board.formation.cho.DefaultChoPlacement;
+import domain.board.formation.cho.InnerHorseChoPlacement;
+import domain.board.formation.cho.LeftInnerHorseChoPlacement;
+import domain.board.formation.cho.RightInnerHorseChoPlacement;
+import domain.board.formation.han.DefaultHanPlacement;
+import domain.board.formation.han.InnerHorseHanPlacement;
+import domain.board.formation.han.LeftInnerHorseHanPlacement;
+import domain.board.formation.han.RightInnerHorseHanPlacement;
 import java.util.Arrays;
 
 public enum PlacementOption {
