+++ src/main/java/domain/board/formation/PlacementStrategy.java
@@ -1,4 +1,4 @@
-package domain.board;
+package domain.board.formation;
 
 import domain.coordination.Coordination;
