+++ src/main/java/domain/board/formation/han/RightInnerHorseHanPlacement.java
@@ -1,4 +1,4 @@
-package domain.board.han;
+package domain.board.formation.han;
 
 import domain.coordination.Coordination;
 import domain.piece.Elephant;
@@ -8,12 +8,12 @@ import domain.piece.Team;
 
 import java.util.Map;
 
-public class InnerHorseHanPlacement extends HanPlacementStrategy {
+public class RightInnerHorseHanPlacement extends HanPlacementStrategy {
 
     @Override
     protected void placeHorseAndElephant(Map<Coordination, Piece> map) {
-        map.put(Coordination.of(2, 1), new Elephant(Team.HAN));
-        map.put(Coordination.of(3, 1), new Horse(Team.HAN));
+        map.put(Coordination.of(2, 1), new Horse(Team.HAN));
+        map.put(Coordination.of(3, 1), new Elephant(Team.HAN));
         map.put(Coordination.of(7, 1), new Horse(Team.HAN));
         map.put(Coordination.of(8, 1), new Elephant(Team.HAN));
     }
